requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.35
asyncpg==0.30.0
asyncssh==2.17.0
paramiko==3.5.0
python-dotenv==1.0.1
python-multipart==0.0.12
pydantic==2.8.2
aiohttp==3.9.1
apscheduler==3.10.4
retrying==1.3.4
aio-pika==9.4.2
aiogram==3.13.1